allow the caller to supply the Client instance
This allows for usecase-specific and/or implementation-specific config. It also makes it easier for clients in OSGi environments, since ClientBuilder.newInstance() depends on the thread context classloader having access to the jax-rs implementation classes, and this isn't always the case in OSGi environments. Such clients can now just use the implementation-specific client builder class directly and not worry about the TCCL.