verify that responses are of the expected media type
Don't try to parse it if it's not the right type.
It's not clear why GR would return a success status that isn't json, but I've seen it happen at least once:
https://rollbar.com/Cru/wsapi/items/578/
It would be helpful to know what the entity body was in this case, but it was lost.
With this change, it won't get lost next time.