nicer error message rendering
In the common case, errors from GR are a simple json payload with a single 'error' key/value pair. For these, let's strip off the json encoding.

For other errors (perhaps ELB-issued ones), the format of the error message is not very predictable, so we print the message as is (but truncated).

This uses a regex for json parsing because this project is intended to be json-parsing-implementation agnostic, and I didn't feel it was appropriate to add this very small parsing task to the Serializer API.