Do not use `release-profile` when releasing
By default, the release plugin activates the maven Super POM's
`release-profile`. [1]
This is irritating, because it runs a second execution of maven-source-plugin,
which results in the source jar being uploaded to artifactory twice during a release.
Artifactory is ok with this if the user is an admin,
but otherwise it doesn't let release artifacts get overwritten,
and the second source jar will be rejected, and the build will fail. [2]

Note: I had originally tried to just rename the ids of the executions in this pom,
as suggested in the first stack overflow answer,
but since we are using a different goal (jar-no-fork instead of jar),
both goals end up in the execution, which isn't what we want.

[1]: https://maven.apache.org/guides/introduction/introduction-to-the-pom.html#Super_POM
[2]: https://stackoverflow.com/q/4251488/1174749